libsubprocess: correct sign-compare in write path
Problem: flux_subprocess_write() compared fbuf_space() (int) against
len (size_t) and passed len to fbuf_write()'s int parameter.

Add explicit casts.

+++ src/common/libsubprocess/subprocess.c
@@ -863,11 +863,11 @@ int flux_subprocess_write (flux_subprocess_t *p,
             log_err ("fbuf_write_watcher_get_buffer");
             return -1;
         }
-        if (fbuf_space (fb) < len) {
+        if ((size_t)fbuf_space (fb) < len) {
             errno = ENOSPC;
             return -1;
         }
-        if ((ret = fbuf_write (fb, buf, len)) < 0) {
+        if ((ret = fbuf_write (fb, buf, (int)len)) < 0) {
             log_err ("fbuf_write");
             return -1;
         }
